requirements/development.txt
-r default.txt

django-extensions==2.2.8
Werkzeug==1.0.0
